+++ common/templates/cookie-consent.d.ts
@@ -1,5 +1,5 @@
 /// <reference path="../../types/index.d.ts" />
 
-declare function renderCookieConsent(options?: Pick<YandexMetrikaOptions, 'pathname'>): string;
+declare function renderCookieConsent(options?: Pick<YandexMetrikaOptions, 'pathname' | 'requestPathname'>): string;
 
 export { renderCookieConsent };